- Completed the new rental feature -with auto completion and toastr notification - includind the api end points

+++ Streams/Dtos/NewRentalDto.cs
@@ -7,7 +7,7 @@ namespace Streams.Dtos
 {
     public class NewRentalDto
     {
-        public byte CustomerId { get; set; }
+        public int CustomerId { get; set; }
         public List<int> MovieIds { get; set; }
     }
 }
